+++ canon-docs/src/app/layout.tsx
@@ -1,8 +1,9 @@
 import type { Metadata } from 'next';
 import { Sidebar } from '../components/Sidebar';
+import { Toolbar } from '@/components/Toolbar';
+import { Providers } from './providers';
 
 import styles from './page.module.css';
-import { Providers } from './providers';
 
 import './globals.css';
 import '/public/core.css';
@@ -19,10 +20,6 @@ export default function RootLayout({
 }: Readonly<{
   children: React.ReactNode;
 }>) {
-  // const cookieStore = await cookies();
-  // const theme = cookieStore.get('theme')?.value || 'light';
-  // const themeName = cookieStore.get('theme-name')?.value || 'default';
-
   return (
     <html
       lang="en"
@@ -34,7 +31,10 @@ export default function RootLayout({
         <Providers>
           <div className={styles.global}>
             <Sidebar />
-            <div className={styles.container}>{children}</div>
+            <div className={styles.container}>
+              <Toolbar />
+              {children}
+            </div>
           </div>
         </Providers>
       </body>
