+++ canon-docs/src/components/Toolbar/nav.module.css
@@ -1,0 +1,84 @@
+.tabs {
+  height: 60px;
+}
+
+.tabsTheme {
+  width: 142px;
+  border-radius: 0.375rem;
+  background-color: var(--canon-surface-2);
+  transition: background-color 0.2s ease-in-out, color 0.2s ease-in-out;
+}
+
+.list {
+  display: flex;
+  position: relative;
+  z-index: 0;
+  gap: 1.5rem;
+}
+
+.tab {
+  height: 60px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  border: 0;
+  margin: 0;
+  outline: 0;
+  background: none;
+  appearance: none;
+  color: var(--canon-text-secondary) !important;
+  font-family: inherit;
+  font-size: 0.75rem;
+  line-height: 1.25rem;
+  font-weight: 500;
+  user-select: none;
+  flex: 1;
+  cursor: pointer;
+  padding: 0;
+
+  &[data-selected] {
+    color: var(--canon-text-primary) !important;
+
+    & p {
+      color: var(--canon-text-primary) !important;
+    }
+  }
+
+  @media (hover: hover) {
+    &:hover {
+      color: var(--canon-text-primary);
+    }
+  }
+
+  &:focus-visible {
+    position: relative;
+
+    &::before {
+      content: '';
+      position: absolute;
+      inset: 0.25rem 0;
+      border-radius: 0.25rem;
+      outline: 2px solid var(--canon-surface-1);
+      outline-offset: -1px;
+    }
+  }
+}
+
+.tab p {
+  color: var(--canon-text-secondary) !important;
+}
+
+.indicator {
+  position: absolute;
+  z-index: -1;
+  left: 0;
+  bottom: -1px;
+  translate: var(--active-tab-left) -50%;
+  width: var(--active-tab-width);
+  height: 1px;
+  border-radius: 0.25rem;
+  background-color: var(--canon-text-primary);
+  transition-property: translate, width, background-color;
+  transition-duration: 200ms;
+  transition-timing-function: ease-in-out;
+}
