+++ canon-docs/src/components/Toolbar/styles.module.css
@@ -1,0 +1,21 @@
+.toolbar {
+  position: sticky;
+  top: 0;
+  left: 0;
+  right: 0;
+  z-index: 1000;
+  background-color: var(--canon-surface-1);
+  border-bottom: 1px solid var(--canon-border-base);
+  height: 60px;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  padding: 0 var(--canon-spacing-md);
+  transition: background-color 0.2s ease-in-out;
+}
+
+.actions {
+  display: flex;
+  align-items: center;
+  gap: var(--canon-spacing-md);
+}
