+++ canon-docs/src/components/Toolbar/tabs.tsx
@@ -1,12 +1,9 @@
 'use client';
 
-import styles from './Tabs.module.css';
 import { Tabs } from '@base-ui-components/react/tabs';
-import { Icon } from '../../../../packages/canon/src/components/Icon';
-import { Text } from '../../../../packages/canon/src/components/Text';
-import { usePathname } from 'next/navigation';
-import { useRouter } from 'next/navigation';
+import { Icon, Text } from '../../../../packages/canon';
 import { usePlayground } from '@/utils/playground-context';
+import styles from './tabs.module.css';
 
 export const TabsVersion = () => {
   const { selectedThemeName, setSelectedThemeName } = usePlayground();
@@ -48,53 +45,6 @@ export const TabsTheme = () => {
         </Tabs.Tab>
         <Tabs.Tab className={styles.tab} value="dark">
           <Icon name="moon" />
-        </Tabs.Tab>
-        <Tabs.Indicator className={styles.indicator} />
-      </Tabs.List>
-    </Tabs.Root>
-  );
-};
-
-export const TabsPages = () => {
-  const pathname = usePathname();
-  const router = useRouter();
-
-  const onValueChange = (value: string) => {
-    if (value === 'docs') {
-      router.push('/');
-    } else {
-      router.push('/playground');
-    }
-  };
-
-  return (
-    <Tabs.Root
-      className={styles.tabs}
-      value={pathname.includes('playground') ? 'playground' : 'docs'}
-      onValueChange={onValueChange}
-    >
-      <Tabs.List className={styles.list}>
-        <Tabs.Tab
-          className={styles.tab}
-          value="docs"
-          onClick={() => {
-            router.push('/');
-          }}
-        >
-          <Text variant="caption" weight="bold">
-            Documentation
-          </Text>
-        </Tabs.Tab>
-        <Tabs.Tab
-          className={styles.tab}
-          value="playground"
-          onClick={() => {
-            router.push('/playground');
-          }}
-        >
-          <Text variant="caption" weight="bold">
-            Playground
-          </Text>
         </Tabs.Tab>
         <Tabs.Indicator className={styles.indicator} />
       </Tabs.List>
